remove reset vector stores endpoint and move to server side updateENV with postUpdate check
